+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * Shows a list of search results in an overlay below the field
   * A result can be selected and customized.
 * A `MacosOverlayFilter` widget can now be used to apply the blurry "frosted glass" effect on surfaces.
+* New widget: `CustomToolbarItem` that enables any widget to be used in the `Toolbar`.
 
 ## [1.0.1]
 * Improvements to the graphical `MacosTimePicker`